se agrego la funcion respectiva a los puntos 3 y 4 las cuales constan de lo mismo y tambien se agregaron dos clases para las balas las cuales ayudan a tener un control sobre ellas

+++ balad.h
@@ -5,9 +5,9 @@
 class balaD
 {
 private:
-    float angulo;
-    float velocidad;
-    float tiempo;
+    int angulo;
+    int velocidad;
+    double tiempo;
 public:
     balaD();
     float getAngulo() const;
